resume_analyzer/requirements.txt
# Add your project dependencies here
pdfplumber
python-docx
spacy
pandas
